+++ examples/react-turnkey-wallet/src/hooks/react-turnkey-wallet/use-pending-unshield.ts
@@ -42,9 +42,5 @@ export function usePendingUnshield(selectedPair: TokenWrapperPairWithMetadata | 
     setPendingUnshieldHash(null);
   }, [selectedPair]);
 
-  return {
-    pendingUnshieldHash,
-    setPendingUnshieldHash,
-    clearStoredPendingUnshield,
-  };
+  return { pendingUnshieldHash, setPendingUnshieldHash, clearStoredPendingUnshield };
 }